Backend/GPS/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.3.0
googlemaps==4.10.0
idna==3.4
numpy==1.24.4
psycopg2==2.9.6
pytz==2023.3
requests==2.31.0
shapely==2.0.1
sqlparse==0.4.4
typing-extensions==4.7.0
urllib3==2.0.3
